fix(core): pass the secure scanf buffer size as unsigned
SHARP_RUNTIME_SCANF_BUFFER handed sscanf_s the size as std::size_t. Microsoft's
CRT reads that argument out of the variadic list as unsigned, so MSVC reports
every %[ and %s in the HTTP-date parser as C4477 and /WX turns it into an
error. The size is now narrowed at the macro; every buffer using it is a small
fixed-size local, so nothing can be lost.

Verified: full build with two jobs and scripts/run_component_tests.sh,
16344/16344 checks across 37 executables.

+++ modules/core/include/SharpRuntime/PortableScan.hpp
@@ -33,10 +33,12 @@
  * @brief Passes a character array to SHARP_RUNTIME_SSCANF with the size `sscanf_s` requires.
  *
  * The argument must be a real array, not a decayed pointer, because the size comes from
- * `sizeof`. It is passed as `std::size_t`, which is what the secure CRT reads it back as.
+ * `sizeof`. The size is narrowed to `unsigned`, which is the type the secure CRT reads back out
+ * of the variadic list -- passing `std::size_t` is what MSVC reports as C4477. Every buffer here
+ * is a small fixed-size local, so the conversion cannot lose anything.
  */
 #if defined(_MSC_VER)
-#  define SHARP_RUNTIME_SCANF_BUFFER(buffer) (buffer), sizeof(buffer)
+#  define SHARP_RUNTIME_SCANF_BUFFER(buffer) (buffer), static_cast<unsigned>(sizeof(buffer))
 #else
 #  define SHARP_RUNTIME_SCANF_BUFFER(buffer) (buffer)
 #endif
